refactor: MVP 패턴 적용을 위한 패키지 정비
