+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.bookingdetail.BookingDetailActivity"
+            android:name=".feature.bookingdetail.view.BookingDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.bookingcomplete.BookingCompleteActivity"
+            android:name=".feature.bookingcomplete.view.BookingCompleteActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.movies.MoviesActivity"
+            android:name=".feature.movies.view.MoviesActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
