+++ app/src/main/java/woowacourse/movie/feature/bookingdetail/view/adapter/DateAdapter.kt
@@ -1,18 +1,19 @@
-package woowacourse.movie.presentation.bookingdetail.adapter
+package woowacourse.movie.feature.bookingdetail.view.adapter
 
+import android.R
 import android.content.Context
 import android.widget.ArrayAdapter
-import woowacourse.movie.presentation.model.MovieDateUiModel
+import woowacourse.movie.feature.model.MovieDateUiModel
 
 class DateAdapter(
     context: Context,
     dates: List<MovieDateUiModel>,
 ) : ArrayAdapter<String>(
         context,
-        android.R.layout.simple_spinner_item,
+        R.layout.simple_spinner_item,
         dates.map { it.toString() },
     ) {
     init {
-        setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+        setDropDownViewResource(R.layout.simple_spinner_dropdown_item)
     }
 }
